pop avant suppression

See merge request cda2021/cfa-portal!47

+++ cfa-front/src/views/Admin/AdminDashboard.vue
@@ -130,9 +130,12 @@ export default {
         );
     },
     deleteUtilisateur(userId) {
-      utilisateurApi.deleteUtilisateur(userId).then(() => this.refreshList());
+      var res = confirm("Êtes-vous sûr de vouloir supprimer?");
+      if(res){
+        utilisateurApi.deleteUtilisateur(userId).then(() => this.refreshList());
+      }
+      
     },
-
   },
 };
 </script>
